refactor: move keystore code
