+++ packages/core/src/key_store/index.ts
@@ -1,0 +1,2 @@
+export { InMemoryKeyStore } from './in_memory_key_store';
+export { KeyStore } from './keystore';
