+++ packages/near-api-js/src/key_stores/keystore.ts
@@ -1,16 +1,1 @@
-import { KeyPair } from '../utils/key_pair';
-
-/**
- * KeyStores are passed to {@link near!Near} via {@link near!NearConfig}
- * and are used by the {@link signer!InMemorySigner} to sign transactions.
- * 
- * @see {@link connect}
- */
-export abstract class KeyStore {
-    abstract setKey(networkId: string, accountId: string, keyPair: KeyPair): Promise<void>;
-    abstract getKey(networkId: string, accountId: string): Promise<KeyPair>;
-    abstract removeKey(networkId: string, accountId: string): Promise<void>;
-    abstract clear(): Promise<void>;
-    abstract getNetworks(): Promise<string[]>;
-    abstract getAccounts(networkId: string): Promise<string[]>;
-}
+export { KeyStore } from '@near-js/core';
